backend/requirements.txt
# API Framework
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.1
python-dotenv>=1.0.0

# LLM Clients
openai>=1.13.0
httpx>=0.27.0

# Agent System
pyautogen>=0.7.5
typing-extensions>=4.9.0

# Database & Auth
supabase>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1

# Utilities
python-multipart>=0.0.9
pyjwt>=2.8.0
loguru>=0.7.2

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0 